--- client/src/components/common/useInfiniteScroll.jsx
@@ -1,271 +1,0 @@
-import { useState, useEffect, useRef, useCallback } from "react";
-
-export const useInfiniteScroll = (
-  loadMoreResults,
-  hasMore,
-  loading,
-  setLoading,
-  delayTime = 10,
-) => {
-  const [userScrolled, setUserScrolled] = useState(false);
-  const [isActionDebouncing, setIsActionDebouncing] = useState(false);
-  const [isShowingLoadingAnimation, setIsShowingLoadingAnimation] =
-    useState(false);
-  const [isLoadLocked, setIsLoadLocked] = useState(false);
-
-  const observer = useRef(null);
-  const lastElementRef = useRef(null);
-  const timerRef = useRef(null);
-  const scrollEndTimerRef = useRef(null);
-  const debugRef = useRef({
-    lastScrollTime: 0,
-    scrollCount: 0,
-    observerCount: 0,
-    loadAttempts: 0,
-    lastLoadTime: 0,
-    hasForceScrollCheck: false,
-  });
-
-  // 관찰자 콜백 함수 - 엘리먼트가 화면에 나타날 때 호출
-  const observerCallback = useCallback(
-    (entries) => {
-      const [entry] = entries;
-
-      if (
-        entry?.isIntersecting &&
-        !loading &&
-        hasMore &&
-        userScrolled &&
-        !isActionDebouncing &&
-        !isLoadLocked
-      ) {
-        loadData("observer");
-      }
-    },
-    [
-      loading,
-      hasMore,
-      userScrolled,
-      isActionDebouncing,
-      isLoadLocked,
-      loadData,
-    ],
-  );
-
-  // 관찰자 초기화 - 의존성 배열에서 hasMore 제거하여 재생성 최소화
-  useEffect(() => {
-    // 기존 옵저버 정리
-    if (observer.current) {
-      observer.current.disconnect();
-    }
-
-    const currentObserver = new IntersectionObserver(observerCallback, {
-      root: null,
-      rootMargin: "0px 0px 2000px 0px", // 2000px 유지
-      threshold: 0.1,
-    });
-
-    observer.current = currentObserver;
-
-    // 마지막 요소가 있다면 즉시 관찰 시작
-    if (lastElementRef.current) {
-      observer.current.observe(lastElementRef.current);
-    }
-
-    return () => {
-      if (currentObserver) {
-        currentObserver.disconnect();
-      }
-
-      if (timerRef.current) {
-        clearTimeout(timerRef.current);
-      }
-
-      if (scrollEndTimerRef.current) {
-        clearTimeout(scrollEndTimerRef.current);
-      }
-    };
-  }, [observerCallback]);
-
-  // 마지막 요소 관찰 설정
-  const lastElementRefCallback = useCallback((node) => {
-    if (!node) {
-      lastElementRef.current = null;
-      return;
-    }
-
-    // 새 노드 관찰 시작
-    lastElementRef.current = node;
-
-    // 옵저버가 이미 존재하면 바로 관찰 시작
-    if (observer.current) {
-      // 관찰 중인 모든 요소 해제
-      observer.current.disconnect();
-      // 새 요소만 관찰
-      observer.current.observe(node);
-    }
-  }, []);
-
-  // 스크롤 감지
-  useEffect(() => {
-    const handleWheel = (e) => {
-      if (e.deltaY > 0 && !userScrolled) {
-        setUserScrolled(true);
-      }
-    };
-
-    const handleTouchMove = () => {
-      if (!userScrolled) {
-        setUserScrolled(true);
-      }
-    };
-
-    // 스크롤 중지 감지 함수
-    const detectScrollEnd = () => {
-      // 이전 타이머 취소
-      if (scrollEndTimerRef.current) {
-        clearTimeout(scrollEndTimerRef.current);
-      }
-
-      // 스크롤 중지 감지를 위한 타이머 설정 (100ms)
-      scrollEndTimerRef.current = setTimeout(() => {
-        // hasMore가 false라도 스크롤 후 처음 한 번은 확인해봄
-        if (!hasMore && !debugRef.current.hasForceScrollCheck) {
-          debugRef.current.hasForceScrollCheck = true;
-          checkScrollPositionForLoad("forceCheck");
-        } else {
-          checkScrollPositionForLoad("scrollEnd");
-        }
-      });
-    };
-
-    // 스크롤 이벤트 (하단 감지용)
-    const handleScroll = () => {
-      // 스크롤 중지 감지
-      detectScrollEnd();
-
-      // 기본 조건 확인 (hasMore 제외 - 필터 변경 직후 강제 체크를 위해)
-      if (
-        loading ||
-        isActionDebouncing ||
-        !userScrolled ||
-        isShowingLoadingAnimation ||
-        isLoadLocked
-      )
-        return;
-
-      const scrollPosition = window.innerHeight + window.scrollY;
-      const scrollHeight = document.body.scrollHeight;
-      const threshold = scrollHeight - 500; // 임계값 늘림 (200px -> 500px)
-
-      if (scrollPosition >= threshold) {
-        // 여기서 hasMore 확인
-        if (!hasMore && !debugRef.current.hasForceScrollCheck) {
-          debugRef.current.hasForceScrollCheck = true;
-        } else if (!hasMore) {
-          return;
-        }
-
-        loadData("scroll");
-      }
-    };
-
-    // 스로틀링된 스크롤 이벤트 핸들러
-    const throttledScrollHandler = () => {
-      const now = Date.now();
-      if (now - debugRef.current.lastScrollTime < 200) return;
-
-      debugRef.current.lastScrollTime = now;
-      handleScroll();
-    };
-
-    window.addEventListener("wheel", handleWheel, { passive: true });
-    window.addEventListener("touchmove", handleTouchMove, { passive: true });
-    window.addEventListener("scroll", throttledScrollHandler, {
-      passive: true,
-    });
-
-    return () => {
-      window.removeEventListener("wheel", handleWheel);
-      window.removeEventListener("touchmove", handleTouchMove);
-      window.removeEventListener("scroll", throttledScrollHandler);
-
-      // 스크롤 종료 타이머 정리
-      if (scrollEndTimerRef.current) {
-        clearTimeout(scrollEndTimerRef.current);
-      }
-    };
-  }, [
-    hasMore,
-    loading,
-    isActionDebouncing,
-    userScrolled,
-    isShowingLoadingAnimation,
-    isLoadLocked,
-    loadData,
-    checkScrollPositionForLoad,
-  ]);
-
-  // 디바운스 액션 - 필터 변경 후 스크롤 이벤트 일시 차단
-  const debounceScrollAction = useCallback(() => {
-    setIsActionDebouncing(true);
-
-    if (timerRef.current) {
-      clearTimeout(timerRef.current);
-    }
-
-    // 필터 변경 시 강제 확인 플래그 초기화
-    debugRef.current.hasForceScrollCheck = false;
-
-    timerRef.current = setTimeout(() => {
-      setIsActionDebouncing(false);
-
-      // 디바운스 종료 후 위치 체크
-      setTimeout(() => {
-        checkScrollPositionForLoad("debounceEnd");
-      }, 100);
-    }, 500); // 디바운스 시간 증가 (400ms -> 500ms)
-  }, [checkScrollPositionForLoad]);
-
-  // 모든 상태 리셋 함수 강화
-  const resetAllStates = useCallback(() => {
-    // 진행 중인 타이머 정리
-    if (timerRef.current) {
-      clearTimeout(timerRef.current);
-      timerRef.current = null;
-    }
-
-    if (scrollEndTimerRef.current) {
-      clearTimeout(scrollEndTimerRef.current);
-      scrollEndTimerRef.current = null;
-    }
-
-    // 필터 변경 시 강제 확인 플래그 초기화 - 매번 새로 시도할 수 있도록
-    debugRef.current.hasForceScrollCheck = false;
-
-    // 모든 상태 초기화
-    setIsShowingLoadingAnimation(false);
-    setIsActionDebouncing(false);
-    setIsLoadLocked(false);
-    setUserScrolled(true); // 스크롤 상태를 true로 강제 설정
-
-    if (setLoading) setLoading(false);
-
-    // 리셋 후 위치 확인
-    setTimeout(() => {
-      // 스크롤 상태 다시 확인하고 필요하면 수동 활성화
-      setUserScrolled(true);
-      window.dispatchEvent(new Event("scroll")); // 스크롤 이벤트 트리거
-      checkScrollPositionForLoad("reset");
-    }, 200);
-  }, [setLoading, checkScrollPositionForLoad]);
-
-  return {
-    lastElementRef: lastElementRefCallback,
-    userScrolled,
-    setUserScrolled,
-    debounceScrollAction,
-    isDelaying: isShowingLoadingAnimation,
-    resetAllStates: resetAllStates,
-  };
-};
