+++ client/src/pages/myPage/_components/common/LoadingIndicator.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+export const LoadingIndicator = () => (
+  <div className="my-6 flex justify-center">
+    <div className="h-10 w-10 animate-spin rounded-full border-t-2 border-b-2 border-indigo-500"></div>
+  </div>
+);
