+++ client/src/pages/myPage/_components/question/settings/hooks.js
@@ -1,28 +1,25 @@
 import { useReducer, useCallback, useEffect } from "react";
 import { ACTIONS, PAGE_SIZE, SCROLL_BATCH_SIZE } from "./constants";
-import { SORT_OPTIONS } from "@/components/common/useFilter";
+import { SORT_OPTIONS } from "@/hooks/useFilter";
 import {
   fetchInterviewsWithFirstQuestion,
   toggleInterviewBookmark,
   batchDeleteInterviews,
 } from "@/api/myPageApi";
-import { loginInfo } from "@/store/loginStore";
 
-// 초기 상태
 const initialState = {
-  results: [], // 전체 결과 (필터링 전)
+  results: [],
   page: 0,
-  visibleResults: [], // 화면에 표시할 결과
+  visibleResults: [],
   hasMore: true,
   loading: false,
   selected: {},
   isDeleteMode: false,
   totalCount: 0,
   error: null,
-  currentSortType: "최신순", // 현재 정렬 타입 저장
+  currentSortType: "최신순",
 };
 
-// 리듀서 함수
 function reducer(state, action) {
   switch (action.type) {
     case ACTIONS.SET_PAGE:
@@ -67,63 +64,43 @@ function reducer(state, action) {
   }
 }
 
-/**
- * 질문 목록 상태 관리 훅 (common 훅 활용 버전)
- */
 export function useQuestionListState() {
   const [state, dispatch] = useReducer(reducer, initialState);
-  const userId = loginInfo((state) => state.userId);
 
-  // 질문 데이터 불러오기
   const fetchQuestionsData = useCallback(
-    async (page = 0, sortType = SORT_OPTIONS.LATEST, userId = null) => {
+    async (page = 0, sortType = SORT_OPTIONS.LATEST) => {
       try {
         dispatch({ type: ACTIONS.SET_LOADING, payload: true });
         dispatch({ type: ACTIONS.SET_ERROR, payload: null });
 
-        // 정렬 타입이 변경되었을 때 처리
         const isFilterChange = sortType !== state.currentSortType;
         const isBookmarkToLatest =
           state.currentSortType === SORT_OPTIONS.BOOKMARK &&
           sortType === SORT_OPTIONS.LATEST;
 
         if (isFilterChange) {
-          // 정렬 타입만 변경 (상태를 초기화하지 않음 - 스크롤 위치 유지)
           dispatch({ type: ACTIONS.SET_SORT_TYPE, payload: sortType });
-
-          // 첫 페이지부터 다시 시작하지만 기존 결과는 유지
           if (page !== 0) {
             page = 0;
           }
         } else {
-          // 정렬 타입 저장
           dispatch({ type: ACTIONS.SET_SORT_TYPE, payload: sortType });
         }
 
-        // API 파라미터 설정
-        const sortBy = "date"; // 항상 날짜순 요청
+        const sortBy = "date";
         const isInitialLoad = page === 0;
+        const batchSize = page === 0 ? SCROLL_BATCH_SIZE * 1 : SCROLL_BATCH_SIZE;
 
-        // 초기 로드 시 더 많은 데이터 불러오기
-        const batchSize =
-          page === 0
-            ? SCROLL_BATCH_SIZE * 1 // 초기 로드 시 2배만 로드 (진짜 무한 스크롤을 위해 줄임)
-            : SCROLL_BATCH_SIZE;
-
-        // API 호출 - userId 전달
-        // 북마크 필터는 모든 항목을 가져오고 클라이언트에서 정렬합니다
         const data = await fetchInterviewsWithFirstQuestion(
           page + 1,
           batchSize,
           sortBy,
-          undefined, // 항상 모든 항목을 가져옵니다
+          undefined,
           isInitialLoad,
-          userId,
         );
 
         if (!data) throw new Error("데이터를 불러올 수 없습니다.");
 
-        // 데이터 가공
         const formattedData = data.map((interview, index) => {
           const firstQuestion = interview.questions?.[0] || null;
           const uniqueId = `${interview.id}-${page}-${index}`;
@@ -142,66 +119,40 @@ export function useQuestionListState() {
               .toISOString()
               .slice(0, 10)
               .replace(/-/g, "."),
-            createdAt: interview.createdAt, // 원본 타임스탬프 저장
+            createdAt: interview.createdAt,
             type: firstQuestion?.type === "PERSONALITY" ? "인성" : "직무",
             bookmarked: interview.bookmarked || false,
             interviewData: interview,
-            userId: interview.userId || userId, // 사용자 ID 저장
-            summary: interview.summary || "", // summary 필드 추가
+            summary: interview.summary || "",
           };
         });
 
-        // 정렬 함수
         const sortResults = (results, type) => {
           if (type === SORT_OPTIONS.BOOKMARK) {
-            // 북마크된 항목이 맨 위에 오도록 정렬 (북마크 우선 정렬)
             return [...results].sort((a, b) => {
-              // 북마크 true가 먼저, false가 나중
               if (a.bookmarked !== b.bookmarked) {
                 return a.bookmarked ? -1 : 1;
               }
-              // 그 다음 최신순
               return new Date(b.createdAt) - new Date(a.createdAt);
             });
           } else {
-            // 최신순 정렬
             return [...results].sort(
               (a, b) => new Date(b.createdAt) - new Date(a.createdAt),
             );
           }
         };
 
-        // 사용자 ID로 필터링 (받은 userId가 있는 경우)
-        const filteredByUser = userId
-          ? formattedData.filter(
-              (item) => !item.userId || item.userId === userId,
-            )
-          : formattedData;
-
-        // 북마크 필터일 때 북마크된 항목만 필터링
-        // 이제 북마크 필터는 필터링이 아닌 정렬 방식만 변경합니다
-        const filteredByBookmark = filteredByUser; // 모든 항목을 유지
-
-        // 데이터 정렬
+        const filteredByUser = formattedData;
+        const filteredByBookmark = filteredByUser;
         const sortedResults = sortResults(filteredByBookmark, sortType);
         const hasMoreData = data.length >= batchSize;
-
-        // hasMoreData가 false여도 필터 변경 후에는 무조건 true로 설정
-        // (최소 1회는 스크롤 동작하게 함)
-        // 북마크 필터에서도 스크롤이 작동하도록 설정
         const adjustedHasMore = isFilterChange || hasMoreData;
 
         if (page === 0) {
-          // 첫 페이지 로드 - 결과 대체
           dispatch({ type: ACTIONS.SET_RESULTS, payload: sortedResults });
 
-          // 필터 변경된 경우 - 모든 전환에서 부드러운 정렬 적용
           if (isFilterChange && state.results.length > 0) {
-            // 기존 데이터를 새 정렬 타입으로 재정렬
-            // sortResults 함수 재사용 (최신순일 때는 이미 날짜 기준으로만 정렬됨)
             const reorderedResults = sortResults(state.results, sortType);
-
-            // 표시 항목 유지하면서 정렬만 변경 - 항목 수 보존
             const visibleCount = Math.max(
               state.visibleResults.length,
               SCROLL_BATCH_SIZE,
@@ -212,7 +163,6 @@ export function useQuestionListState() {
               payload: reorderedResults.slice(0, visibleCount),
             });
           } else {
-            // 첫 로드 시 표시할 아이템 수 (진짜 무한 스크롤을 위해 줄임)
             const initialVisibleCount = Math.min(
               sortedResults.length,
               SCROLL_BATCH_SIZE,
@@ -224,19 +174,15 @@ export function useQuestionListState() {
             });
           }
         } else {
-          // 추가 페이지 로드 - 결과 병합
-          // 중복 제거된 새 결과 배열 생성
           const uniqueResults = mergeUniqueResults(
             state.results,
             sortedResults,
           );
 
-          // 정렬 적용
           const sortedMergedResults = sortResults(uniqueResults, sortType);
 
           dispatch({ type: ACTIONS.SET_RESULTS, payload: sortedMergedResults });
 
-          // 스크롤 시 추가 표시할 항목 수 계산 (누적 페이지 기준으로 자름)
           const visibleCount = (page + 1) * SCROLL_BATCH_SIZE;
 
           dispatch({
@@ -249,7 +195,6 @@ export function useQuestionListState() {
         dispatch({ type: ACTIONS.SET_PAGE, payload: page });
         dispatch({ type: ACTIONS.SET_TOTAL_COUNT, payload: data.length });
       } catch (error) {
-        console.error("[로그] 데이터 로드 오류:", error);
         dispatch({
           type: ACTIONS.SET_ERROR,
           payload: `데이터 로드 중 오류가 발생했습니다.`,
@@ -261,13 +206,11 @@ export function useQuestionListState() {
     [state.results, state.currentSortType, state.visibleResults.length],
   );
 
-  // 북마크 토글 함수
   const toggleQuestionBookmark = useCallback(
-    async (id, userId = null) => {
+    async (id) => {
       try {
         dispatch({ type: ACTIONS.SET_LOADING, payload: true });
 
-        // 항목 찾기
         const targetItem = state.results.find((q) => q.id === id);
         if (!targetItem) {
           throw new Error("북마크할 항목을 찾을 수 없습니다.");
@@ -276,10 +219,8 @@ export function useQuestionListState() {
         const originalId = targetItem.originalId;
         const newBookmarkState = !targetItem.bookmarked;
 
-        // API 호출 - userId 전달
-        await toggleInterviewBookmark(originalId, newBookmarkState, userId);
+        await toggleInterviewBookmark(originalId, newBookmarkState);
 
-        // 전체 결과 배열에서 북마크 상태 업데이트
         const updatedResults = state.results.map((item) =>
           item.originalId === originalId
             ? { ...item, bookmarked: newBookmarkState }
@@ -288,11 +229,8 @@ export function useQuestionListState() {
 
         dispatch({ type: ACTIONS.SET_RESULTS, payload: updatedResults });
 
-        // 북마크 필터 모드일 때 특수 처리
         if (state.currentSortType === SORT_OPTIONS.BOOKMARK) {
-          // 북마크 해제 시: 해당 항목 제거 후 재정렬
           if (!newBookmarkState) {
-            // originalId가 일치하는 항목을 제외한 나머지만 표시
             const filteredVisible = state.visibleResults.filter(
               (item) => item.originalId !== originalId,
             );
@@ -301,7 +239,6 @@ export function useQuestionListState() {
               payload: filteredVisible,
             });
           } else {
-            // 북마크 추가 시: 전체 재정렬 (북마크된 항목이 앞으로 오도록)
             const sortedResults = [...updatedResults].sort((a, b) => {
               if (a.bookmarked !== b.bookmarked) {
                 return a.bookmarked ? -1 : 1;
@@ -309,7 +246,6 @@ export function useQuestionListState() {
               return new Date(b.createdAt) - new Date(a.createdAt);
             });
 
-            // 현재 표시 중인 항목 수 유지
             const currentCount = state.visibleResults.length;
             dispatch({
               type: ACTIONS.SET_VISIBLE_RESULTS,
@@ -317,7 +253,6 @@ export function useQuestionListState() {
             });
           }
         } else {
-          // 일반 모드에서는 표시 중인 항목들의 북마크 상태만 업데이트
           const updatedVisible = state.visibleResults.map((item) =>
             item.originalId === originalId
               ? { ...item, bookmarked: newBookmarkState }
@@ -340,12 +275,10 @@ export function useQuestionListState() {
     [state.results, state.currentSortType, state.visibleResults],
   );
 
-  // 항목 선택 토글
   const toggleSelectItem = useCallback((id) => {
     dispatch({ type: ACTIONS.TOGGLE_SELECT, payload: id });
   }, []);
 
-  // 삭제 모드 토글
   const toggleDeleteMode = useCallback(() => {
     if (state.isDeleteMode) {
       dispatch({ type: ACTIONS.CLEAR_SELECTED });
@@ -353,16 +286,14 @@ export function useQuestionListState() {
     dispatch({ type: ACTIONS.TOGGLE_DELETE_MODE });
   }, [state.isDeleteMode]);
 
-  // 선택된 항목 삭제
   const markAsDeleted = useCallback(
-    async (selectedItems, userId = null) => {
+    async (selectedItems) => {
       try {
         const itemsToDelete = Object.keys(selectedItems).filter(
           (id) => selectedItems[id],
         );
         if (itemsToDelete.length === 0) return;
 
-        // UI 업데이트
         const updatedResults = state.results.filter(
           (item) => !itemsToDelete.includes(item.id),
         );
@@ -376,13 +307,12 @@ export function useQuestionListState() {
         dispatch({ type: ACTIONS.CLEAR_SELECTED });
         dispatch({ type: ACTIONS.TOGGLE_DELETE_MODE });
 
-        // 서버 업데이트
         const interviewIds = itemsToDelete
           .map((id) => state.results.find((q) => q.id === id)?.interviewId)
           .filter((id) => id);
 
         if (interviewIds.length > 0) {
-          await batchDeleteInterviews(interviewIds, userId);
+          await batchDeleteInterviews(interviewIds);
         }
       } catch (error) {
         dispatch({
@@ -394,21 +324,16 @@ export function useQuestionListState() {
     [state.results],
   );
 
-  // 중복 없이 결과 병합 (개선된 버전)
   function mergeUniqueResults(existingResults, newResults) {
-    // ID 기반 맵 생성
     const resultMap = new Map();
 
-    // 기존 결과 추가
     existingResults.forEach((item) => {
       resultMap.set(item.originalId, item);
     });
 
-    // 새 결과 추가 (중복 시 북마크 상태 유지)
     newResults.forEach((item) => {
       const existing = resultMap.get(item.originalId);
       if (existing) {
-        // 기존 항목의 북마크 상태 유지하면서 업데이트
         resultMap.set(item.originalId, {
           ...item,
           bookmarked: existing.bookmarked,
@@ -418,7 +343,6 @@ export function useQuestionListState() {
       }
     });
 
-    // 맵을 배열로 변환
     return Array.from(resultMap.values());
   }
