+++ server/src/controllers/mypage.controller.js
@@ -7,51 +7,36 @@ import bcrypt from "bcrypt";
 
 export const getMyBookmarks = async (req, res) => {
   try {
-    // 쿼리 파라미터에서 userId를 가져오거나, 로그인된 사용자 ID 사용
     const userId = req.user.userId;
-
-    // 페이지네이션 파라미터
     const page = parseInt(req.query.page, 10) || 1;
     const pageSize = parseInt(req.query.pageSize, 10) || 10;
     const skip = (page - 1) * pageSize;
 
-    // userId가 없으면 401 에러 반환
-    // if (!userId) {
-    //   return res.status(401).json({ message: "인증이 필요합니다." });
-    // }
-
-    // 기본 필터링 조건 - 사용자 ID 및 북마크 상태
     const where = {
       userId,
       bookmarked: true,
     };
 
     try {
-      // 직군/직무 필터
       if (req.query.role) {
         where.interview = {
           role: req.query.role,
         };
       }
 
-      // 질문 유형 필터
       if (req.query.type) {
         where.type = req.query.type === "직무" ? "JOB" : "PERSONALITY";
       }
     } catch (filterError) {
-      console.error("[서버] 필터 적용 중 오류:", filterError);
-      // 필터 오류 발생 시 기본 조건만 사용
       where.interview = undefined;
       where.type = undefined;
     }
 
     try {
-      // 총 질문 수 조회 (페이지네이션용)
       const totalCount = await prisma.question.count({
         where,
       });
 
-      // 현재 페이지 데이터 조회
       const questions = await prisma.question.findMany({
         where,
         include: {
@@ -62,7 +47,6 @@ export const getMyBookmarks = async (req, res) => {
         take: pageSize,
       });
 
-      // 페이지네이션 정보 포함하여 응답
       res.status(200).json({
         questions,
         totalCount,
@@ -71,14 +55,12 @@ export const getMyBookmarks = async (req, res) => {
         totalPages: Math.ceil(totalCount / pageSize),
       });
     } catch (dbError) {
-      console.error("[서버] 데이터베이스 쿼리 오류:", dbError);
       return res.status(500).json({
         message: "데이터 조회 중 오류가 발생했습니다.",
         error: dbError.message,
       });
     }
   } catch (error) {
-    console.error("북마크 조회 오류:", error);
     res.status(500).json({ message: "서버 오류가 발생했습니다." });
   }
 };
@@ -86,10 +68,8 @@ export const getMyBookmarks = async (req, res) => {
 // 사용자 정보 조회
 export const getUserInfo = async (req, res) => {
   try {
-    // 쿼리 파라미터에서 userId를 가져오기
     const clientUserId = req.user.userId;
 
-    // 사용자 정보 조회
     const user = await prisma.user.findUnique({
       where: { id: clientUserId },
     });
@@ -98,12 +78,10 @@ export const getUserInfo = async (req, res) => {
       return res.status(404).json({ message: "사용자를 찾을 수 없습니다." });
     }
 
-    // 비밀번호는 제외하고 반환
     const { password, ...userInfo } = user;
 
     res.status(200).json(userInfo);
   } catch (error) {
-    console.error("[서버] 사용자 정보 조회 오류:", error);
     res.status(500).json({ message: "서버 오류가 발생했습니다." });
   }
 };
@@ -111,44 +89,31 @@ export const getUserInfo = async (req, res) => {
 // 사용자 정보 업데이트
 export const updateUserInfo = async (req, res) => {
   try {
-    // 토큰에서 사용자 ID 가져오기
     const userId = req.user.userId;
-
-    // if (!userId) {
-    //   console.log("[서버] 인증 실패: 토큰에서 사용자 ID를 찾을 수 없음");
-    //   return res.status(401).json({ message: "인증이 필요합니다." });
-    // }
-
     const { password, role, career } = req.body;
 
-    // 업데이트할 데이터 객체 생성
     const updateData = {};
 
-    // 비밀번호가 제공된 경우 해싱
     if (password) {
       const salt = await bcrypt.genSalt(10);
       const hashedPassword = await bcrypt.hash(password, salt);
       updateData.password = hashedPassword;
     }
 
-    // 직무가 제공된 경우 업데이트
     if (role !== undefined) {
       updateData.role = role;
     }
 
-    // 경력이 제공된 경우 업데이트
     if (career !== undefined) {
       updateData.career = career;
     }
 
-    // 업데이트할 데이터가 없는 경우
     if (Object.keys(updateData).length === 0) {
       return res
         .status(400)
         .json({ message: "업데이트할 정보가 제공되지 않았습니다." });
     }
 
-    // 먼저 사용자가 존재하는지 확인
     const existingUser = await prisma.user.findUnique({
       where: { id: userId },
     });
@@ -158,13 +123,11 @@ export const updateUserInfo = async (req, res) => {
     }
 
     try {
-      // 사용자 정보 업데이트
       const updatedUser = await prisma.user.update({
         where: { id: userId },
         data: updateData,
       });
 
-      // 비밀번호는 제외하고 반환
       const { password: _, ...userInfo } = updatedUser;
 
       return res.status(200).json({
@@ -172,11 +135,9 @@ export const updateUserInfo = async (req, res) => {
         user: userInfo,
       });
     } catch (updateError) {
-      console.error("[서버] Prisma 업데이트 에러:", updateError);
       throw updateError;
     }
   } catch (error) {
-    console.error("[서버] 사용자 정보 업데이트 오류:", error);
     return res.status(500).json({ message: "서버 오류가 발생했습니다." });
   }
 };
@@ -184,19 +145,9 @@ export const updateUserInfo = async (req, res) => {
 // 회원 탈퇴
 export const deleteUserAccount = async (req, res) => {
   try {
-    // 토큰에서 사용자 ID 가져오기
     const userId = req.user.userId;
-    console.log("[서버] 토큰에서 가져온 사용자 ID:", userId);
-
-    // if (!userId) {
-    //   console.log("[서버] 인증 실패: 토큰에서 사용자 ID를 찾을 수 없음");
-    //   return res.status(401).json({ message: "인증이 필요합니다." });
-    // }
-
-    // 비밀번호 확인
     const password = req.body?.password;
 
-    // 먼저 사용자가 존재하는지 확인
     const existingUser = await prisma.user.findUnique({
       where: { id: userId },
     });
@@ -205,7 +156,6 @@ export const deleteUserAccount = async (req, res) => {
       return res.status(404).json({ message: "사용자를 찾을 수 없습니다." });
     }
 
-    // 비밀번호 확인이 제공된 경우 검증
     if (password) {
       const isPasswordValid = await bcrypt.compare(
         password,
