+++ server/src/controllers/question.controller.js
@@ -4,40 +4,27 @@ import * as questionService from "../services/question.service.js";
 // 모든 질문 조회
 export const getAllQuestions = async (req, res) => {
   try {
-    // 쿼리 파라미터에서 userId를 가져오거나, 로그인된 사용자 ID 사용
     const userId = req.query.userId || req.user?.id;
-
-    // userId가 없으면 401 에러 반환
-    // if (!userId) {
-    //   return res.status(401).json({ message: "인증이 필요합니다." });
-    // }
-
-    // 쿼리 파라미터에서 필터와 페이지네이션 정보 추출
     const { page, pageSize, sortBy, bookmarked } = req.query;
 
-    // 페이지네이션 설정
     const pagination = {
       page: page ? parseInt(page) : 1,
       pageSize: pageSize ? parseInt(pageSize) : 10,
     };
 
-    // 필터 설정
     const filters = {
       sortBy: sortBy || "date",
       bookmarked: bookmarked === "true",
     };
 
-    // 서비스에 사용자 ID, 페이지네이션, 필터 정보 전달
     const result = await questionService.getAllQuestions(
       userId,
       pagination,
       filters
     );
 
-    // 기존 API와의 호환성을 위해 questions 배열만 반환
     res.status(200).json(result.questions);
   } catch (error) {
-    console.error("질문 조회 오류:", error);
     res.status(500).json({ message: "서버 오류가 발생했습니다." });
   }
 };
@@ -46,10 +33,6 @@ export const getAllQuestions = async (req, res) => {
 export const getQuestionById = async (req, res) => {
   try {
     const userId = req.user.id;
-
-    // if (!userId) {
-    //   return res.status(401).json({ message: "인증이 필요합니다." });
-    // }
 
     const question = await questionService.getQuestionById(id);
 
@@ -123,7 +106,6 @@ export const toggleBookmark = async (req, res) => {
     const { userId } = req.user;
     const authenticatedUserId = userId || req.user?.id;
 
-    // userId가 없으면 401 에러 반환
     if (!authenticatedUserId) {
       return res.status(401).json({ message: "인증이 필요합니다." });
     }
